general-ui-autotests/requirements.txt
pytest <= 4.1.1
attrs <= 19.1.0
pytest-html
requests
Appium-Python-Client<=0.46
selenium==3.141.0
babel